+++ src/components/PostDetails.tsx
@@ -9,9 +9,9 @@ import { CommentData } from '../types/Comment';
 import { useAppDispatch, useAppSelector } from '../app/hooks';
 import {
   createNewComment,
-  deletedComment,
   init as fetchComment,
 } from '../features/commentsSlice';
+import { Comments } from './Comments';
 
 type Props = {
   post: Post;
@@ -50,16 +50,6 @@ export const PostDetails: React.FC<Props> = ({ post }) => {
     }
   };
 
-  const deleteComment = async (commentId: number) => {
-    // we delete the comment immediately so as
-    // not to make the user wait long for the actual deletion
-
-    dispatch({ type: 'comments/deleteComments', payload: commentId });
-    dispatch(deletedComment(commentId));
-
-    await commentsApi.deleteComment(commentId);
-  };
-
   return (
     <div className="content" data-cy="PostDetails">
       <div className="block">
@@ -90,37 +80,7 @@ export const PostDetails: React.FC<Props> = ({ post }) => {
         )}
 
         {!loaded && !hasError && comments.length > 0 && (
-          <>
-            <p className="title is-4">Comments:</p>
-
-            {comments.map(comment => (
-              <article
-                className="message is-small"
-                key={comment.id}
-                data-cy="Comment"
-              >
-                <div className="message-header">
-                  <a href={`mailto:${comment.email}`} data-cy="CommentAuthor">
-                    {comment.name}
-                  </a>
-
-                  <button
-                    data-cy="CommentDelete"
-                    type="button"
-                    className="delete is-small"
-                    aria-label="delete"
-                    onClick={() => deleteComment(comment.id)}
-                  >
-                    delete button
-                  </button>
-                </div>
-
-                <div className="message-body" data-cy="CommentBody">
-                  {comment.body}
-                </div>
-              </article>
-            ))}
-          </>
+          <Comments />
         )}
 
         {!loaded && !hasError && !visible && (
