+++ src/components/PostsList.tsx
@@ -1,19 +1,12 @@
-import classNames from 'classnames';
 import React from 'react';
 import { Post } from '../types/Post';
-import { useAppDispatch, useAppSelector } from '../app/hooks';
+import { Posts } from './Posts';
 
 type Props = {
   posts: Post[],
 };
 
-export const PostsList: React.FC<Props> = ({
-  posts,
-}) => {
-  const dispatch = useAppDispatch();
-  const selectedPostId = useAppSelector((state) => state
-    .selectedPost.selectedPost?.id);
-
+export const PostsList: React.FC<Props> = () => {
   return (
     <div data-cy="PostsList">
       <p className="title">Posts:</p>
@@ -27,35 +20,7 @@ export const PostsList: React.FC<Props> = ({
           </tr>
         </thead>
 
-        <tbody>
-          {posts.map(post => (
-            <tr key={post.id} data-cy="Post">
-              <td data-cy="PostId">{post.id}</td>
-              <td data-cy="PostTitle">{post.title}</td>
-              <td className="has-text-right is-vcentered">
-                <button
-                  type="button"
-                  data-cy="PostButton"
-                  className={classNames(
-                    'button',
-                    'is-link',
-                    {
-                      'is-light': post.id !== selectedPostId,
-                    },
-                  )}
-                  onClick={() => {
-                    dispatch({
-                      type: 'selectedPost/setSelectedPost',
-                      payload: post.id === selectedPostId ? null : post,
-                    });
-                  }}
-                >
-                  {post.id === selectedPostId ? 'Close' : 'Open'}
-                </button>
-              </td>
-            </tr>
-          ))}
-        </tbody>
+        <Posts />
       </table>
     </div>
   );
