+++ src/components/UserSelector.tsx
@@ -1,31 +1,20 @@
 import React, { useEffect, useState } from 'react';
 import classNames from 'classnames';
-import { setAuthor } from '../features/authorSlice';
-import { useAppDispatch, useAppSelector } from '../app/hooks';
-// import { init } from '../features/userSlice';
+import { useAppSelector } from '../app/hooks';
+import { Users } from './Users';
 
 export const UserSelector: React.FC = () => {
-  // `users` are loaded from the API, so for the performance reasons
-  // we load them once in the `UsersContext` when the `App` is opened
-  // and now we can easily reuse the `UserSelector` in any form
-  const users = useAppSelector(
-    (state) => state.users.users,
-  );
   const selectedUser = useAppSelector(
     (state) => state.author.author,
   );
   const [expanded, setExpanded] = useState(false);
-  const dispatch = useAppDispatch();
 
   useEffect(() => {
     if (!expanded) {
       return;
     }
 
-    // we save a link to remove the listener later
     const handleDocumentClick = () => {
-      // we close the Dropdown on any click (inside or outside)
-      // So there is not need to check if we clicked inside the list
       setExpanded(false);
     };
 
@@ -35,8 +24,6 @@ export const UserSelector: React.FC = () => {
     return () => {
       document.removeEventListener('click', handleDocumentClick);
     };
-  // we don't want to listening for outside clicks
-  // when the Dopdown is closed
   }, [expanded]);
 
   return (
@@ -66,22 +53,7 @@ export const UserSelector: React.FC = () => {
       </div>
 
       <div className="dropdown-menu" id="dropdown-menu" role="menu">
-        <div className="dropdown-content">
-          {users.map(user => (
-            <a
-              key={user.id}
-              href={`#user-${user.id}`}
-              onClick={() => {
-                dispatch(setAuthor(user));
-              }}
-              className={classNames('dropdown-item', {
-                'is-active': user.id === selectedUser?.id,
-              })}
-            >
-              {user.name}
-            </a>
-          ))}
-        </div>
+        <Users />
       </div>
     </div>
   );
